requirements.txt
# Python dependencies

confluent-kafka==1.3.0
python-dotenv==0.19.2